Revert "[WIP] Creating tops"
